tgbot: major change

+++ tg_bot/modules/msg_deleting.py
@@ -7,7 +7,7 @@ from telegram.ext import CommandHandler, Filters
 from telegram.ext.dispatcher import run_async
 from telegram.utils.helpers import mention_html
 
-from tg_bot import dispatcher, LOGGER
+from tg_bot import dispatcher, LOGGER, SUDO_USERS
 from tg_bot.modules.helper_funcs.chat_status import user_admin, can_delete
 from tg_bot.modules.log_channel import loggable
 
@@ -17,9 +17,15 @@ from tg_bot.modules.log_channel import loggable
 @loggable
 def purge(bot: Bot, update: Update, args: List[str]) -> str:
     msg = update.effective_message  # type: Optional[Message]
+    chat = update.effective_chat  # type: Optional[Chat]
+    user = update.effective_user  # type: Optional[User]
+
+    admin = chat.get_member(int(user.id))
+    if ( not admin.can_delete_messages ) and ( not int(user.id) in SUDO_USERS ):
+        update.effective_message.reply_text("You don't have sufficient permissions to delete messages!")
+        return ""
+
     if msg.reply_to_message:
-        user = update.effective_user  # type: Optional[User]
-        chat = update.effective_chat  # type: Optional[Chat]
         if can_delete(chat, bot.id):
             message_id = msg.reply_to_message.message_id
             delete_to = msg.message_id - 1
@@ -73,6 +79,12 @@ def del_message(bot: Bot, update: Update) -> str:
     if update.effective_message.reply_to_message:
         user = update.effective_user  # type: Optional[User]
         chat = update.effective_chat  # type: Optional[Chat]
+        
+        admin = chat.get_member(int(user.id))
+        if ( not admin.can_delete_messages ) and ( not int(user.id) in SUDO_USERS ):
+            update.effective_message.reply_text("You don't have sufficient permissions to delete messages!")
+            return ""
+
         if can_delete(chat, bot.id):
             update.effective_message.reply_to_message.delete()
             update.effective_message.delete()
